small fixes

+++ src/main/java/administration.java
@@ -1,7 +1,7 @@
 
 public class administration extends user implements java.io.Serializable{
     
-    administration(String id,String name,String surname,String password,String address,String email,int phone){
+    administration(String id,String name,String surname,String password,String address,String email,long phone){
         super(id,name,surname,password,address,email,phone);
     }
